wip - AbstractGeneratorOperation
