+++ src/Operation/Implode.php
@@ -16,7 +16,7 @@ use loophp\collection\Contract\Operation;
  *
  * @implements Operation<TKey, T>
  */
-final class Implode extends AbstractGeneratorOperation implements Operation
+final class Implode extends AbstractOperation implements Operation
 {
     public function __construct(string $glue)
     {
